Add unit tests to for the python code from the notebook.
Extracting and cleaning the code cells from the notebook
to allow some Python unit tests.

Some things are undefined outside of the notebook
environment and need to be tweaked.

Extensive library mocking is used for the part that runs
as a script. Methods are better for testing.

Added license headers in tools scripts.
Added unittest.sh script and travis entry.